Move mkdirp into libi3
